Fix a few bugs (video player showing offline, showing video/topic available links, etc); pdb crept in), reorganize code
