+++ kalite/templates/topic.html
@@ -1,6 +1,6 @@
 {% extends 'base_distributed.html' %}
-
 {% load i18n %}
+{% load my_filters %}
 
 {% block watch_active %}active{% endblock watch_active %}
 
@@ -21,14 +21,14 @@
 
 
 {% block headjs %}
-    <script type="text/javascript">
+<script type="text/javascript">
     $(function() {
         $('.video-unavailable').removeAttr("href"); 
 
         // TODO(bcipolli): use javascript to translate here
         $('a.video-unavailable').attr('title', '{% trans "This video is not downloaded." %}');
     });
-    </script>
+</script>
 {% endblock headjs %}
 
 
@@ -43,12 +43,16 @@
     {% for subtopic in topics %}
     <li>
         <span>
-        {% if backup_vids_available and subtopic.nvideos_local != subtopic.nvideos_known %}
-            <a href="{{ subtopic.path }}" class="topic-available client-online-only" title="{{subtopic.nvideos_known}}/{{subtopic.nvideos_known}} {% trans "video(s) are available." %}">
-                {% trans subtopic.title %}
-            </a>
-        {% endif %}
-        <a href="{{ subtopic.path }}" class="{% if backup_vids_available and subtopic.nvideos_local != subtopic.nvideos_known %}not-client-online-only {% endif %}topic-{% if subtopic.nvideos_local == 0 %}un{% endif %}available" title="{{subtopic.nvideos_local}}/{{subtopic.nvideos_known}} {% trans "video(s) are downloaded." %}">
+        {% comment %}Four scenarios:
+        1. Offline, local video
+        2. Offline, no local video
+        3. Online, local video
+        4. Online, no local video
+        {% endcomment %}
+        <a href="{{ subtopic.path }}" class="topic-link client-online-only topic-{% if not subtopic.nvideos_local and not backup_vids_available %}un{% endif %}available" title="{{ subtopic.nvideos_local }}/{{ subtopic.nvideos_known }} {% trans "video(s) are available locally." %}">
+            {% trans subtopic.title %}
+        </a>
+        <a href="{{ subtopic.path }}" class="topic-link not-client-online-only topic-{% if subtopic.nvideos_local == 0 %}un{% endif %}available" title="{{ subtopic.nvideos_local }}/{{subtopic.nvideos_known}} {% trans "video(s) are available locally." %}">
             {% trans subtopic.title %}
         </a>
         </span>
@@ -65,15 +69,16 @@
     <li>
         <!--<a href="mplayer:///khan/videofiles/{{ video.youtube_id }}.ogv">{{ video.title }}</a>-->
         <span class="progress-circle" data-youtube-id="{{ video.youtube_id }}"></span>
-        {% if backup_vids_available and subtopic.nvideos_local != subtopic.nvideos_known %}
-            <a class="video-link video-available client-online-only" href="v/{{ video.readable_id }}/">
-                {% trans video.title %}
-            </a>
-        {% endif %}
-        <a href="v/{{ video.readable_id }}/" class="{% if backup_vids_available and not video.on_disk %}client-online-only {% endif %}video-available" title="{{subtopic.nvideos_local}}/{{subtopic.nvideos_known}} {% trans "video(s) are downloaded." %}">
+        {% comment %}Four scenarios:
+        1. Offline, local video
+        2. Offline, no local video
+        3. Online, local video
+        4. Online, no local video
+        {% endcomment %}
+        <a href="v/{{ video.readable_id }}/" class="video-link client-online-only video-{% if not video.on_disk and not backup_vids_available %}un{% endif %}available">
             {% trans video.title %}
         </a>
-        <a href="v/{{ video.readable_id }}/" class="video-link {% if backup_vids_available and not video.on_disk %} not-client-online-only {% endif %}video-{% if not video.on_disk and not backup_vids_available %}un{% endif %}available">
+        <a href="v/{{ video.readable_id }}/" class="video-link not-client-online-only video-{% if not video.on_disk %}un{% endif %}available">
             {% trans video.title %}
         </a>
     </li>
