+++ kalite/templates/video.html
@@ -76,28 +76,25 @@
 {% endif %}
 <div class="clear"></div>
 
-{% if video_exists %}
-    {% if use_mplayer %}
+{% if use_mplayer %}
 
-        <h2 id="video-player">
-            <a href="" id="launch_mplayer">
-                <img class="video-thumb" src="{{ video.thumbnail_url }}" style="width: 100%;" title="Click to play!">
-                <br/><br/>
-                Launch video in mplayer!
-            </a>
-        </h2>
+    <h2 id="video-player" {% if not video.on_disk %} class="{% if not backup_vids_available %}not-{% endif %}client-online-only"{% endif %}>
+        <a href="" id="launch_mplayer">
+            <img class="video-thumb" src="{{ video.thumbnail_url }}" style="width: 100%;" title="Click to play!">
+            <br/><br/>
+            Launch video in mplayer!
+        </a>
+    </h2>
 
-    {% else %}
+{% else %}
 
-        <div id="video-player" {% if not video.ondisk %} class="client-online-only"{% endif %}>
-            <video id="video-element" class="video-js vjs-default-skin" controls preload="auto" width="100%" height="600" {% if video.thumbnail_url %}poster="{{ video.thumbnail_url }}" {% endif %}data-setup="{}">
-                <source src="{{ video.stream_url }}" type="{{ video.stream_type }}" />
-                {% if video.subtitles_url %}<track kind="captions" src="{{ video.subtitles_url }}" srclang="en" label="On" />{% endif %}
-            </video>
-            {% if video.thumbnail_url %}<img class="video-thumb" src="{{ video.thumbnail_url }}" style="position: absolute; left: -9999999px;">{% endif %}
-        </div>
-
-    {% endif %}
+    <div id="video-player" {% if not video.on_disk %} class="{% if not backup_vids_available %}not-{% endif %}client-online-only"{% endif %}>
+        <video id="video-element" class="video-js vjs-default-skin" controls preload="auto" width="100%" height="600" {% if video.thumbnail_url %}poster="{{ video.thumbnail_url }}" {% endif %}data-setup="{}">
+            <source src="{{ video.stream_url }}" type="{{ video.stream_type }}" />
+            {% if video.subtitles_url %}<track kind="captions" src="{{ video.subtitles_url }}" srclang="en" label="On" />{% endif %}
+        </video>
+        {% if video.thumbnail_url %}<img class="video-thumb" src="{{ video.thumbnail_url }}" style="position: absolute; left: -9999999px;">{% endif %}
+    </div>
 
 {% endif %}
